+++ Sources/RichTextKit/Context/RichTextContext.swift
@@ -7,21 +7,24 @@
 //
 
 import SwiftUI
+import Combine
 
 /**
  This observable context can be used to affect and observe a
  ``RichTextEditor`` and its content.
 
+ Use ``handle(_:)`` to trigger a change, e.g. to change font,
+ text style, alignment, select a range, etc. You can observe
+ the various published properties to keep your UI updated as
+ the context changes.
+
  The SwiftUI ``RichTextEditor`` uses this context as well as
- a ``RichTextCoordinator`` to keep itself updated when state
- changes. The context also has properties and functions that
- are used to set context state via the coordinator.
+ a ``RichTextCoordinator`` to keep itself updated.
  */
 public class RichTextContext: ObservableObject {
 
     /// Create a new rich text context instance.
     public init() {}
-
 
     // MARK: - Not yet observable properties
 
@@ -30,56 +33,28 @@ public class RichTextContext: ObservableObject {
 
      Note that the property is read-only and not `@Published`
      to avoid redrawing the editor when it changes, which is
-     done constantly as the user types. We should find a way
-     to observe this property without this happening.
-
-     The best way to observe this property is to use the raw
-     `text` binding that you pass into your text editor. The
-     editor will however not redraw if you change this value
-     from the outside, since it aims to avoid costly redraws.
+     done as the user types. We should find a way to observe
+     it without this happening. The best way to observe this
+     property is to use the raw `text` binding that you pass
+     into the text editor. The editor will not redraw if you
+     change this value from the outside.
 
      Until then, use `setAttributedString(to:)` to change it.
      */
     public internal(set) var attributedString = NSAttributedString()
 
-    /**
-     The currently selected range, if any.
-
-     Note that the property is read-only and not `@Published`
-     to avoid redrawing the editor when it changes, which is
-     done constantly as the user types. We should find a way
-     to observe this property without this happening.
-
-     Until then, use ``selectRange(_:)`` to change the value.
-     */
+    /// The currently selected range, if any.
     public internal(set) var selectedRange = NSRange()
 
+    // MARK: - Public properies
 
-    // MARK: - Observable properies
-
-    /// The current background color, if any.
+    /// Whether or not the text is currently being edited.
     @Published
-    public var backgroundColor: ColorRepresentable?
+    public var isEditingText = false
 
-    /// Whether or not the current rich text can be copied.
+    /// The current text alignment, if any.
     @Published
-    public var canCopy = false
-
-    /// Whether or not the latest undone change can be redone.
-    @Published
-    public var canRedoLatestChange = false
-
-    /// Whether or not the latest change can be undone.
-    @Published
-    public var canUndoLatestChange = false
-
-    /// Whether or not the indent level can be decreased.
-    @Published
-    public var canDecreaseIndent = true
-
-    /// Whether or not the indent level can be increased.
-    @Published
-    public var canIncreaseIndent = true
+    public var textAlignment: RichTextAlignment = .left
 
     /// The current font name.
     @Published
@@ -89,80 +64,73 @@ public class RichTextContext: ObservableObject {
     @Published
     public var fontSize = CGFloat.standardRichTextFontSize
 
+    /// The currently highlighted range, if any.
+    public var highlightedRange: NSRange?
+
+    /// Use this Publisher to emit any attribute changes to textView.
+    public let userActionPublisher: PassthroughSubject<RichTextAction, Never> = .init()
+
+    // MARK: - Internal properties
+
+    /// The current background color, if any.
+    @Published
+    public internal(set) var backgroundColor: ColorRepresentable?
+
+    /// Whether or not the current rich text can be copied.
+    @Published
+    public internal(set) var canCopy = false
+
+    /// Whether or not the latest undone change can be redone.
+    @Published
+    public internal(set) var canRedoLatestChange = false
+
+    /// Whether or not the latest change can be undone.
+    @Published
+    public internal(set) var canUndoLatestChange = false
+
+    /// Whether or not the indent level can be decreased.
+    @Published
+    public internal(set) var canDecreaseIndent = true
+
+    /// Whether or not the indent level can be increased.
+    @Published
+    public internal(set) var canIncreaseIndent = true
+
     /// The current foreground color, if any.
     @Published
-    public var foregroundColor: ColorRepresentable?
-
-    /// The currently highlighted range, if any.
-    @Published
-    public var highlightedRange: NSRange?
+    public internal(set) var foregroundColor: ColorRepresentable?
 
     /// The style to apply when highlighting a range.
     @Published
-    public var highlightingStyle = RichTextHighlightingStyle.standard
+    public internal(set) var highlightingStyle = RichTextHighlightingStyle.standard
 
     /// Whether or not the current text is bold.
     @Published
-    public var isBold = false
-
-    /// Whether or not the text is currently being edited.
-    @Published
-    public var isEditingText = false
+    public internal(set) var isBold = false
 
     /// Whether or not the current text is italic.
     @Published
-    public var isItalic = false
+    public internal(set) var isItalic = false
 
     /// Whether or not the current text is striked through.
     @Published
-    public var isStrikethrough = false
+    public internal(set) var isStrikethrough = false
 
     /// Whether or not the current text is underlined.
     @Published
-    public var isUnderlined = false
+    public internal(set) var isUnderlined = false
 
     /// The current strikethrough color, if any.
     @Published
-    public var strikethroughColor: ColorRepresentable?
+    public internal(set) var strikethroughColor: ColorRepresentable?
 
     /// The current stroke color, if any.
     @Published
-    public var strokeColor: ColorRepresentable?
-
-    /// The current text alignment, if any.
-    @Published
-    public var textAlignment: RichTextAlignment = .left
+    public internal(set) var strokeColor: ColorRepresentable?
 
     /// The current underline color, if any.
     @Published
-    public var underlineColor: ColorRepresentable?
-
-
-    // MARK: - Internal trigger properties
-
-    /// Set this property to trigger a certain action.
-    @Published
-    var triggerAction: RichTextAction?
-
-    /// Set this property to trigger a paste operation.
-    @Published
-    var shouldPasteImage: (image: ImageRepresentable, atIndex: Int, moveCursor: Bool)?
-
-    /// Set this property to trigger an image paste operation.
-    @Published
-    var shouldPasteImages: (images: [ImageRepresentable], atIndex: Int, moveCursor: Bool)?
-
-    /// Set this property to trigger a text paste operation.
-    @Published
-    var shouldPasteText: (text: String, atIndex: Int, moveCursor: Bool)?
-
-    /// Set this property to trigger a string update.
-    @Published
-    var shouldSetAttributedString: NSAttributedString?
-
-    /// Set this property to trigger a range change.
-    @Published
-    var shouldSelectRange = NSRange()
+    public internal(set) var underlineColor: ColorRepresentable?
 }
 
 public extension RichTextContext {
@@ -182,6 +150,7 @@ public extension RichTextContext {
 
     /// Set ``highlightedRange`` to a new, optional range.
     func highlightRange(_ range: NSRange?) {
+        userActionPublisher.send(.setHighlightedRange(range))
         highlightedRange = range
     }
 
@@ -192,7 +161,15 @@ public extension RichTextContext {
         moveCursorToPastedContent: Bool = false
     ) {
         let index = index ?? selectedRange.location
-        shouldPasteImage = (image, index, moveCursorToPastedContent)
+        userActionPublisher.send(
+            .pasteImage(
+                RichTextInsertion(
+                    content: image,
+                    at: index,
+                    moveCursor: moveCursorToPastedContent
+                )
+            )
+        )
     }
 
     /// Paste images into the editor, at a certain index.
@@ -202,7 +179,15 @@ public extension RichTextContext {
         moveCursorToPastedContent: Bool = false
     ) {
         let index = index ?? selectedRange.location
-        shouldPasteImages = (images, index, moveCursorToPastedContent)
+        userActionPublisher.send(
+            .pasteImages(
+                RichTextInsertion(
+                    content: images,
+                    at: index,
+                    moveCursor: moveCursorToPastedContent
+                )
+            )
+        )
     }
 
     /// Paste text into the editor, at a certain index.
@@ -212,7 +197,15 @@ public extension RichTextContext {
         moveCursorToPastedContent: Bool = false
     ) {
         let index = index ?? selectedRange.location
-        shouldPasteText = (text, index, moveCursorToPastedContent)
+        userActionPublisher.send(
+            .pasteText(
+                RichTextInsertion(
+                    content: text,
+                    at: index,
+                    moveCursor: moveCursorToPastedContent
+                )
+            )
+        )
     }
 
     /// Reset the attributed string.
@@ -234,7 +227,7 @@ public extension RichTextContext {
     /// Set a new range and start editing.
     func selectRange(_ range: NSRange) {
         isEditingText = true
-        shouldSelectRange = range
+        userActionPublisher.send(.selectRange(range))
     }
 
     /// Set the attributed string to a new plain text.
@@ -246,7 +239,7 @@ public extension RichTextContext {
     func setAttributedString(to string: NSAttributedString) {
         let mutable = NSMutableAttributedString(attributedString: string)
         mutable.setRichTextFontSize(fontSize, at: mutable.richTextRange)
-        shouldSetAttributedString = mutable
+        userActionPublisher.send(.setAttributedString(mutable))
     }
 
     /// Set ``isEditingText`` to `false`.