+++ Sources/RichTextKit/Coordinator/RichTextCoordinator+Subscriptions.swift
@@ -11,29 +11,16 @@ import SwiftUI
 
 extension RichTextCoordinator {
 
-    /// Make the coordinator subscribe to context changes.
-    func subscribeToContextChanges() {
+    func subscribeToUserActions() {
+        richTextContext.userActionPublisher.sink { [weak self] action in
+            self?.handle(action)
+        }
+        .store(in: &cancellables)
+
         subscribeToAlignment()
-        subscribeToBackgroundColor()
         subscribeToFontName()
         subscribeToFontSize()
-        subscribeToForegroundColor()
-        subscribeToHighlightedRange()
-        subscribeToHighlightingStyle()
-        subscribeToIsBold()
         subscribeToIsEditingText()
-        subscribeToIsItalic()
-        subscribeToIsStrikethrough()
-        subscribeToIsUnderlined()
-        subscribeToShouldPasteImage()
-        subscribeToShouldPasteImages()
-        subscribeToShouldPasteText()
-        subscribeToShouldSelectRange()
-        subscribeToShouldSetAttributedString()
-        subscribeToStrikethroughColor()
-        subscribeToStrokeColor()
-        subscribeToTriggerAction()
-        subscribeToUnderlineColor()
     }
 }
 
@@ -42,15 +29,32 @@ private extension RichTextCoordinator {
     func handle(_ action: RichTextAction?) {
         guard let action else { return }
         switch action {
-        case .copy:
-            textView.copySelection()
+        case .copy: textView.copySelection()
         case .dismissKeyboard:
             textView.resignFirstResponder()
+        case .pasteImage(let image):
+            pasteImage(image)
+        case .pasteImages(let images):
+            pasteImages(images)
+        case .pasteText(let text):
+            pasteText(text)
         case .print: break
         case .redoLatestChange:
             textView.redoLatestChange()
             syncContextWithTextView()
+        case .selectRange(let range):
+            setSelectedRange(to: range)
         case .setAlignment: break
+        case .setAttributedString(let string):
+            setAttributedString(to: string)
+        case .setColor(let color, let newValue):
+            setColor(color, to: newValue)
+        case .setHighlightedRange(let range):
+            setHighlightedRange(to: range)
+        case .setHighlightingStyle(let style):
+            textView.highlightingStyle = style
+        case .setStyle(let style, let newValue):
+            setStyle(style, to: newValue)
         case .stepFontSize: break
         case .stepIndent(let points):
             textView.stepRichTextIndent(points: points)
@@ -62,248 +66,72 @@ private extension RichTextCoordinator {
         }
     }
 
-    func subscribeToTriggerAction() {
-        richTextContext.$triggerAction
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.handle($0)
-                })
-            .store(in: &cancellables)
-    }
-
-
     func subscribeToAlignment() {
         richTextContext.$textAlignment
             .sink(
-                receiveCompletion: { _ in },
                 receiveValue: { [weak self] in
                     self?.textView.setRichTextAlignment($0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToBackgroundColor() {
-        richTextContext.$backgroundColor
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    guard let color = $0 else { return }
-                    self?.textView.setRichTextColor(.background, to: color)
-                })
+                }
+            )
             .store(in: &cancellables)
     }
 
     func subscribeToFontName() {
         richTextContext.$fontName
             .sink(
-                receiveCompletion: { _ in },
                 receiveValue: { [weak self] in
                     self?.textView.setRichTextFontName($0)
-                })
+                }
+            )
             .store(in: &cancellables)
     }
 
     func subscribeToFontSize() {
         richTextContext.$fontSize
             .sink(
-                receiveCompletion: { _ in },
                 receiveValue: { [weak self] in
                     self?.textView.setRichTextFontSize($0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToForegroundColor() {
-        richTextContext.$foregroundColor
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    guard let color = $0 else { return }
-                    self?.textView.setRichTextColor(.foreground, to: color)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToHighlightedRange() {
-        richTextContext.$highlightedRange
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setHighlightedRange(to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToHighlightingStyle() {
-        richTextContext.$highlightingStyle
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.textView.highlightingStyle = $0
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToIsBold() {
-        richTextContext.$isBold
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setStyle(.bold, to: $0)
-                })
+                }
+            )
             .store(in: &cancellables)
     }
 
     func subscribeToIsEditingText() {
         richTextContext.$isEditingText
             .sink(
-                receiveCompletion: { _ in },
                 receiveValue: { [weak self] in
                     self?.setIsEditing(to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToIsItalic() {
-        richTextContext.$isItalic
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setStyle(.italic, to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToIsStrikethrough() {
-        richTextContext.$isStrikethrough
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setStyle(.strikethrough, to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToIsUnderlined() {
-        richTextContext.$isUnderlined
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setStyle(.underlined, to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToShouldPasteImage() {
-        richTextContext.$shouldPasteImage
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.pasteImage($0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToShouldPasteImages() {
-        richTextContext.$shouldPasteImages
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.pasteImages($0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToShouldPasteText() {
-        richTextContext.$shouldPasteText
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.pasteText($0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToShouldSetAttributedString() {
-        richTextContext.$shouldSetAttributedString
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setAttributedString(to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToShouldSelectRange() {
-        richTextContext.$shouldSelectRange
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    self?.setSelectedRange(to: $0)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToStrokeColor() {
-        richTextContext.$strokeColor
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    guard let color = $0 else { return }
-                    self?.textView.setRichTextColor(.stroke, to: color)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToStrikethroughColor() {
-        richTextContext.$strikethroughColor
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    guard let color = $0 else { return }
-                    self?.textView.setRichTextColor(.strikethrough, to: color)
-                })
-            .store(in: &cancellables)
-    }
-
-    func subscribeToUnderlineColor() {
-        richTextContext.$underlineColor
-            .sink(
-                receiveCompletion: { _ in },
-                receiveValue: { [weak self] in
-                    guard let color = $0 else { return }
-                    self?.textView.setRichTextColor(.underline, to: color)
-                })
+                }
+            )
             .store(in: &cancellables)
     }
 }
 
-internal extension RichTextCoordinator {
+extension RichTextCoordinator {
 
-    func pasteImage(_ data: (image: ImageRepresentable, atIndex: Int, moveCursor: Bool)?) {
+    func pasteImage(_ data: RichTextInsertion<ImageRepresentable>?) {
         guard let data = data else { return }
         textView.pasteImage(
-            data.image,
-            at: data.atIndex,
+            data.content,
+            at: data.at,
             moveCursorToPastedContent: data.moveCursor
         )
     }
 
-    func pasteImages(_ data: (images: [ImageRepresentable], atIndex: Int, moveCursor: Bool)?) {
+    func pasteImages(_ data: RichTextInsertion<[ImageRepresentable]>?) {
         guard let data = data else { return }
         textView.pasteImages(
-            data.images,
-            at: data.atIndex,
+            data.content,
+            at: data.at,
             moveCursorToPastedContent: data.moveCursor
         )
     }
 
-    func pasteText(_ data: (text: String, atIndex: Int, moveCursor: Bool)?) {
+    func pasteText(_ data: RichTextInsertion<String>?) {
         guard let data = data else { return }
         textView.pasteText(
-            data.text,
-            at: data.atIndex,
+            data.content,
+            at: data.at,
             moveCursorToPastedContent: data.moveCursor
         )
     }
@@ -363,7 +191,7 @@ internal extension RichTextCoordinator {
 extension ColorRepresentable {
 
     #if iOS || os(tvOS) || os(visionOS)
-    static var textColor: ColorRepresentable { .label }
+    public static var textColor: ColorRepresentable { .label }
     #endif
 }
 #endif