+++ Sources/RichTextKit/RichTextView_AppKit.swift
@@ -30,12 +30,14 @@ open class RichTextView: NSTextView, RichTextViewComponent {
     /// The configuration to use by the rich text view.
     public var configuration: Configuration = .standard
 
+    /// The theme for coloring and setting style to text view.
+    public var theme: Theme = .standard
+
     /// The style to use when highlighting text in the view.
     public var highlightingStyle: RichTextHighlightingStyle = .standard
 
     /// The image configuration to use by the rich text view.
     public var imageConfiguration: RichTextImageConfiguration = .disabled
-
 
     // MARK: - Overrides
 
@@ -93,7 +95,6 @@ open class RichTextView: NSTextView, RichTextViewComponent {
             .scrollWheel(with: event)
     }
 
-
     // MARK: - Setup
 
     /**
@@ -109,19 +110,27 @@ open class RichTextView: NSTextView, RichTextViewComponent {
         format: RichTextDataFormat
     ) {
         attributedString = .empty
-        setupInitialFontSize()
         attributedString = text
         allowsImageEditing = true
         allowsUndo = true
-        backgroundColor = .clear
-        trySetupInitialTextColor(for: text) {
-            textColor = .textColor
-        }
         imageConfiguration = standardImageConfiguration(for: format)
         layoutManager?.defaultAttachmentScaling = NSImageScaling.scaleProportionallyDown
         setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
+        setupConfiguration()
+        setupTheme()
     }
 
+    // MARK: - Internal
+
+    func setupConfiguration() {
+        isContinuousSpellCheckingEnabled = configuration.isContinuousSpellCheckingEnabled
+    }
+
+    func setupTheme() {
+        font = theme.font
+        textColor = theme.fontColor
+        backgroundColor = theme.backgroundColor
+    }
 
     // MARK: - Open Functionality
 
@@ -172,7 +181,6 @@ open class RichTextView: NSTextView, RichTextViewComponent {
     }
 }
 
-
 // MARK: - Public Extensions
 
 public extension RichTextView {
@@ -189,7 +197,6 @@ public extension RichTextView {
     }
 }
 
-
 // MARK: - RichTextProvider
 
 public extension RichTextView {
@@ -205,7 +212,6 @@ public extension RichTextView {
         window?.firstResponder == self
     }
 }
-
 
 // MARK: - RichTextWriter