+++ Sources/RichTextKit/RichTextView_UIKit.swift
@@ -26,7 +26,6 @@ extension RichTextView: UIDropInteractionDelegate {}
  */
 open class RichTextView: UITextView, RichTextViewComponent {
 
-
     // MARK: - Initializers
 
     public convenience init(
@@ -45,13 +44,21 @@ open class RichTextView: UITextView, RichTextViewComponent {
         self.setup(with: string, format: format)
     }
 
-
     // MARK: - Properties
 
     /// The configuration to use by the rich text view.
     public var configuration: Configuration = .standard {
         didSet {
             isScrollEnabled = configuration.isScrollingEnabled
+            allowsEditingTextAttributes = configuration.allowsEditingTextAttributes
+            autocapitalizationType = configuration.autocapitalizationType
+            spellCheckingType = configuration.spellCheckingType
+        }
+    }
+
+    public var theme: Theme = .standard {
+        didSet {
+            setupTheme()
         }
     }
 
@@ -74,7 +81,6 @@ open class RichTextView: UITextView, RichTextViewComponent {
         }
     }
 
-
     #if iOS || os(visionOS)
 
     /// The image drop interaction to use.
@@ -89,13 +95,11 @@ open class RichTextView: UITextView, RichTextViewComponent {
 
     #endif
 
-
     /// Keeps track of the first time a valid frame is set.
     private var isInitialFrameSetupNeeded = true
 
     /// Keeps track of the data format used by the view.
     private var richTextDataFormat: RichTextDataFormat = .archivedData
-
 
     // MARK: - Overrides
 
@@ -144,7 +148,6 @@ open class RichTextView: UITextView, RichTextViewComponent {
     }
     #endif
 
-
     // MARK: - Setup
 
     /**
@@ -160,21 +163,23 @@ open class RichTextView: UITextView, RichTextViewComponent {
         format: RichTextDataFormat
     ) {
         attributedString = .empty
-        setupInitialFontSize()
         imageConfiguration = standardImageConfiguration(for: format)
         text.autosizeImageAttachments(maxSize: imageAttachmentMaxSize)
         attributedString = text
-        allowsEditingTextAttributes = false
-        autocapitalizationType = .sentences
-        backgroundColor = .clear
         richTextDataFormat = format
-        spellCheckingType = .no
-        trySetupInitialTextColor(for: text) {
-            textColor = .label
-        }
         setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
+        setupTheme()
     }
 
+    // MARK: - Internal functionality
+
+    func setupTheme() {
+        if text.isEmpty {
+            font = theme.font
+            textColor = theme.fontColor
+        }
+        backgroundColor = theme.backgroundColor
+    }
 
     // MARK: - Open Functionality
 
@@ -245,7 +250,6 @@ open class RichTextView: UITextView, RichTextViewComponent {
         undoManager?.undo()
     }
 
-
     #if iOS || os(visionOS)
 
     // MARK: - UIDropInteractionDelegate
@@ -295,7 +299,6 @@ open class RichTextView: UITextView, RichTextViewComponent {
         performImageDrop(with: session, at: range)
         performTextDrop(with: session, at: range)
     }
-
 
     // MARK: - Drop Interaction Support
 
@@ -357,7 +360,6 @@ private extension UIDropSession {
 }
 #endif
 
-
 // MARK: - Public Extensions
 
 public extension RichTextView {
@@ -385,7 +387,6 @@ public extension RichTextView {
     }
 }
 
-
 // MARK: - RichTextProvider
 
 public extension RichTextView {
@@ -396,7 +397,6 @@ public extension RichTextView {
         set { attributedText = newValue }
     }
 }
-
 
 // MARK: - RichTextWriter